+++ card_test.go
@@ -8,7 +8,7 @@ import (
 
 func TestRankIntegers(t *testing.T) {
 	for i := 0; i < 13; i++ {
-		assert.Equal(t, i, intRanks[i])
+		assert.Equal(t, int32(i), intRanks[i])
 	}
 }
 
@@ -22,23 +22,23 @@ func TestString(t *testing.T) {
 }
 
 func TestBitRank(t *testing.T) {
-	assert.Equal(t, 2048, NewCard("Ks").BitRank())
+	assert.Equal(t, int32(2048), NewCard("Ks").BitRank())
 }
 
 func TestPrimeProductFromHand(t *testing.T) {
-	assert.Equal(t, 20387, primeProductFromHand([]Card{8398611, 134236965, 33564957}))
-	assert.Equal(t, 61161, primeProductFromHand([]Card{8398611, 134236965, 33564957, 135427}))
-	assert.Equal(t, 183483, primeProductFromHand([]Card{8398611, 134236965, 33564957, 135427, 139523}))
+	assert.Equal(t, int32(20387), primeProductFromHand([]Card{8398611, 134236965, 33564957}))
+	assert.Equal(t, int32(61161), primeProductFromHand([]Card{8398611, 134236965, 33564957, 135427}))
+	assert.Equal(t, int32(183483), primeProductFromHand([]Card{8398611, 134236965, 33564957, 135427, 139523}))
 }
 
 func TestPrimeProductFromRankBits(t *testing.T) {
-	assert.Equal(t, 2, primeProductFromRankBits(1))
-	assert.Equal(t, 42, primeProductFromRankBits(11))
-	assert.Equal(t, 110, primeProductFromRankBits(21))
-	assert.Equal(t, 2310, primeProductFromRankBits(31))
-	assert.Equal(t, 4290, primeProductFromRankBits(55))
-	assert.Equal(t, 1785, primeProductFromRankBits(78))
-	assert.Equal(t, 1326, primeProductFromRankBits(99))
-	assert.Equal(t, 34034, primeProductFromRankBits(121))
-	assert.Equal(t, 30107, primeProductFromRankBits(344))
+	assert.Equal(t, int32(2), primeProductFromRankBits(1))
+	assert.Equal(t, int32(42), primeProductFromRankBits(11))
+	assert.Equal(t, int32(110), primeProductFromRankBits(21))
+	assert.Equal(t, int32(2310), primeProductFromRankBits(31))
+	assert.Equal(t, int32(4290), primeProductFromRankBits(55))
+	assert.Equal(t, int32(1785), primeProductFromRankBits(78))
+	assert.Equal(t, int32(1326), primeProductFromRankBits(99))
+	assert.Equal(t, int32(34034), primeProductFromRankBits(121))
+	assert.Equal(t, int32(30107), primeProductFromRankBits(344))
 }
